Frontend is now independent from platform-specific resolvers
- symbol_resolver relies solely on module metadata structures, not on image_info;
- modules_loaded event now supplies instance_ids ony.

+++ test-helpers/helpers_win32.cpp
@@ -49,7 +49,7 @@ namespace micro_profiler
 		unsigned image::get_symbol_rva(const char *name) const
 		{
 			if (void *symbol = ::GetProcAddress(static_cast<HMODULE>(get()), name))
-				return static_cast<byte *>(symbol) - load_address_ptr();
+				return static_cast<unsigned>(static_cast<byte *>(symbol) - load_address_ptr());
 			throw runtime_error("Symbol specified was not found!");
 		}
 	}
